+++ RELEASE-NOTES.md
@@ -3,6 +3,7 @@
 * Update to bouncy-castle 1.68
 * Added database type "atomic_file" for use with MrPlow
 * Added data migration tool for MrPlow data
+* On non-64-bit JVM default to use non-TLS seeds (Issue #179)
 
 ## 1.8.0